factor out common export types
